feat(bidi): Add tool executor
