+++ src/strands/experimental/hooks/__init__.py
@@ -1,10 +1,20 @@
-"""Experimental hook functionality that has not yet reached stability."""
+"""Experimental hook functionality that has not yet reached stability.
+
+BidiAgent hooks are also available here to avoid circular imports.
+"""
 
 from .events import (
     AfterModelInvocationEvent,
     AfterToolInvocationEvent,
     BeforeModelInvocationEvent,
     BeforeToolInvocationEvent,
+    BidiAfterInvocationEvent,
+    BidiAfterToolCallEvent,
+    BidiAgentInitializedEvent,
+    BidiBeforeInvocationEvent,
+    BidiBeforeToolCallEvent,
+    BidiInterruptionEvent,
+    BidiMessageAddedEvent,
 )
 
 __all__ = [
@@ -12,4 +22,12 @@ __all__ = [
     "AfterToolInvocationEvent",
     "BeforeModelInvocationEvent",
     "AfterModelInvocationEvent",
+    # BidiAgent hooks
+    "BidiAgentInitializedEvent",
+    "BidiBeforeInvocationEvent",
+    "BidiAfterInvocationEvent",
+    "BidiMessageAddedEvent",
+    "BidiBeforeToolCallEvent",
+    "BidiAfterToolCallEvent",
+    "BidiInterruptionEvent",
 ]
