+++ src/strands/tools/executors/_executor.py
@@ -7,10 +7,11 @@ thread pools, etc.).
 import abc
 import logging
 import time
-from typing import TYPE_CHECKING, Any, AsyncGenerator, cast
+from typing import TYPE_CHECKING, Any, AsyncGenerator, Union, cast
 
 from opentelemetry import trace as trace_api
 
+from ...experimental.hooks.events import BidiAfterToolCallEvent, BidiBeforeToolCallEvent
 from ...hooks import AfterToolCallEvent, BeforeToolCallEvent
 from ...telemetry.metrics import Trace
 from ...telemetry.tracer import get_tracer, serialize
@@ -21,6 +22,7 @@ from ..structured_output._structured_output_context import StructuredOutputConte
 
 if TYPE_CHECKING:  # pragma: no cover
     from ...agent import Agent
+    from ...experimental.bidi.agent.agent import BidiAgent
 
 logger = logging.getLogger(__name__)
 
@@ -29,8 +31,82 @@ class ToolExecutor(abc.ABC):
     """Abstract base class for tool executors."""
 
     @staticmethod
+    def _is_bidi_agent(agent: Union["Agent", "BidiAgent"]) -> bool:
+        """Check if the agent is a BidiAgent instance.
+        
+        Uses isinstance() with runtime import to avoid circular imports.
+        """
+        # Import at runtime to avoid circular dependency
+        from ...experimental.bidi.agent.agent import BidiAgent
+        
+        return isinstance(agent, BidiAgent)
+
+    @staticmethod
+    async def _invoke_before_tool_call_hook(
+        agent: Union["Agent", "BidiAgent"],
+        tool_func: Any,
+        tool_use: ToolUse,
+        invocation_state: dict[str, Any],
+    ) -> tuple[Any, list]:
+        """Invoke the appropriate before tool call hook based on agent type."""
+        if ToolExecutor._is_bidi_agent(agent):
+            return await agent.hooks.invoke_callbacks_async(
+                BidiBeforeToolCallEvent(
+                    agent=agent,
+                    selected_tool=tool_func,
+                    tool_use=tool_use,
+                    invocation_state=invocation_state,
+                )
+            )
+        else:
+            return await agent.hooks.invoke_callbacks_async(
+                BeforeToolCallEvent(
+                    agent=agent,
+                    selected_tool=tool_func,
+                    tool_use=tool_use,
+                    invocation_state=invocation_state,
+                )
+            )
+
+    @staticmethod
+    async def _invoke_after_tool_call_hook(
+        agent: Union["Agent", "BidiAgent"],
+        selected_tool: Any,
+        tool_use: ToolUse,
+        invocation_state: dict[str, Any],
+        result: ToolResult,
+        exception: Exception | None = None,
+        cancel_message: str | None = None,
+    ) -> tuple[Any, list]:
+        """Invoke the appropriate after tool call hook based on agent type."""
+        if ToolExecutor._is_bidi_agent(agent):
+            return await agent.hooks.invoke_callbacks_async(
+                BidiAfterToolCallEvent(
+                    agent=agent,
+                    selected_tool=selected_tool,
+                    tool_use=tool_use,
+                    invocation_state=invocation_state,
+                    result=result,
+                    exception=exception,
+                    cancel_message=cancel_message,
+                )
+            )
+        else:
+            return await agent.hooks.invoke_callbacks_async(
+                AfterToolCallEvent(
+                    agent=agent,
+                    selected_tool=selected_tool,
+                    tool_use=tool_use,
+                    invocation_state=invocation_state,
+                    result=result,
+                    exception=exception,
+                    cancel_message=cancel_message,
+                )
+            )
+
+    @staticmethod
     async def _stream(
-        agent: "Agent",
+        agent: "Agent | BidiAgent",
         tool_use: ToolUse,
         tool_results: list[ToolResult],
         invocation_state: dict[str, Any],
@@ -48,7 +124,7 @@ class ToolExecutor(abc.ABC):
         - Interrupt handling for human-in-the-loop workflows
 
         Args:
-            agent: The agent for which the tool is being executed.
+            agent: The agent (Agent or BidiAgent) for which the tool is being executed.
             tool_use: Metadata and inputs for the tool to be executed.
             tool_results: List of tool results from each tool execution.
             invocation_state: Context for the tool invocation.
@@ -85,13 +161,9 @@ class ToolExecutor(abc.ABC):
             }
         )
 
-        before_event, interrupts = await agent.hooks.invoke_callbacks_async(
-            BeforeToolCallEvent(
-                agent=agent,
-                selected_tool=tool_func,
-                tool_use=tool_use,
-                invocation_state=invocation_state,
-            )
+        # Invoke appropriate before tool call hook based on agent type
+        before_event, interrupts = await ToolExecutor._invoke_before_tool_call_hook(
+            agent, tool_func, tool_use, invocation_state
         )
 
         if interrupts:
@@ -109,15 +181,9 @@ class ToolExecutor(abc.ABC):
                 "status": "error",
                 "content": [{"text": cancel_message}],
             }
-            after_event, _ = await agent.hooks.invoke_callbacks_async(
-                AfterToolCallEvent(
-                    agent=agent,
-                    tool_use=tool_use,
-                    invocation_state=invocation_state,
-                    selected_tool=None,
-                    result=cancel_result,
-                    cancel_message=cancel_message,
-                )
+            
+            after_event, _ = await ToolExecutor._invoke_after_tool_call_hook(
+                agent, None, tool_use, invocation_state, cancel_result, cancel_message=cancel_message
             )
             yield ToolResultEvent(after_event.result)
             tool_results.append(after_event.result)
@@ -147,14 +213,9 @@ class ToolExecutor(abc.ABC):
                     "status": "error",
                     "content": [{"text": f"Unknown tool: {tool_name}"}],
                 }
-                after_event, _ = await agent.hooks.invoke_callbacks_async(
-                    AfterToolCallEvent(
-                        agent=agent,
-                        selected_tool=selected_tool,
-                        tool_use=tool_use,
-                        invocation_state=invocation_state,
-                        result=result,
-                    )
+                
+                after_event, _ = await ToolExecutor._invoke_after_tool_call_hook(
+                    agent, selected_tool, tool_use, invocation_state, result
                 )
                 yield ToolResultEvent(after_event.result)
                 tool_results.append(after_event.result)
@@ -184,14 +245,8 @@ class ToolExecutor(abc.ABC):
 
             result = cast(ToolResult, event)
 
-            after_event, _ = await agent.hooks.invoke_callbacks_async(
-                AfterToolCallEvent(
-                    agent=agent,
-                    selected_tool=selected_tool,
-                    tool_use=tool_use,
-                    invocation_state=invocation_state,
-                    result=result,
-                )
+            after_event, _ = await ToolExecutor._invoke_after_tool_call_hook(
+                agent, selected_tool, tool_use, invocation_state, result
             )
 
             yield ToolResultEvent(after_event.result)
@@ -204,22 +259,16 @@ class ToolExecutor(abc.ABC):
                 "status": "error",
                 "content": [{"text": f"Error: {str(e)}"}],
             }
-            after_event, _ = await agent.hooks.invoke_callbacks_async(
-                AfterToolCallEvent(
-                    agent=agent,
-                    selected_tool=selected_tool,
-                    tool_use=tool_use,
-                    invocation_state=invocation_state,
-                    result=error_result,
-                    exception=e,
-                )
+            
+            after_event, _ = await ToolExecutor._invoke_after_tool_call_hook(
+                agent, selected_tool, tool_use, invocation_state, error_result, exception=e
             )
             yield ToolResultEvent(after_event.result)
             tool_results.append(after_event.result)
 
     @staticmethod
     async def _stream_with_trace(
-        agent: "Agent",
+        agent: Union["Agent", "BidiAgent"],
         tool_use: ToolUse,
         tool_results: list[ToolResult],
         cycle_trace: Trace,
@@ -231,7 +280,7 @@ class ToolExecutor(abc.ABC):
         """Execute tool with tracing and metrics collection.
 
         Args:
-            agent: The agent for which the tool is being executed.
+            agent: The agent (Agent or BidiAgent) for which the tool is being executed.
             tool_use: Metadata and inputs for the tool to be executed.
             tool_results: List of tool results from each tool execution.
             cycle_trace: Trace object for the current event loop cycle.
@@ -277,7 +326,7 @@ class ToolExecutor(abc.ABC):
     # pragma: no cover
     def _execute(
         self,
-        agent: "Agent",
+        agent: Union["Agent", "BidiAgent"],
         tool_uses: list[ToolUse],
         tool_results: list[ToolResult],
         cycle_trace: Trace,
@@ -288,7 +337,7 @@ class ToolExecutor(abc.ABC):
         """Execute the given tools according to this executor's strategy.
 
         Args:
-            agent: The agent for which tools are being executed.
+            agent: The agent (Agent or BidiAgent) for which tools are being executed.
             tool_uses: Metadata and inputs for the tools to be executed.
             tool_results: List of tool results from each tool execution.
             cycle_trace: Trace object for the current event loop cycle.
