+++ src/strands/tools/executors/concurrent.py
@@ -1,7 +1,7 @@
 """Concurrent tool executor implementation."""
 
 import asyncio
-from typing import TYPE_CHECKING, Any, AsyncGenerator
+from typing import TYPE_CHECKING, Any, AsyncGenerator, Union
 
 from typing_extensions import override
 
@@ -12,6 +12,7 @@ from ._executor import ToolExecutor
 
 if TYPE_CHECKING:  # pragma: no cover
     from ...agent import Agent
+    from ...experimental.bidi.agent.agent import BidiAgent
     from ..structured_output._structured_output_context import StructuredOutputContext
 
 
@@ -21,7 +22,7 @@ class ConcurrentToolExecutor(ToolExecutor):
     @override
     async def _execute(
         self,
-        agent: "Agent",
+        agent: Union["Agent", "BidiAgent"],
         tool_uses: list[ToolUse],
         tool_results: list[ToolResult],
         cycle_trace: Trace,
@@ -32,7 +33,7 @@ class ConcurrentToolExecutor(ToolExecutor):
         """Execute tools concurrently.
 
         Args:
-            agent: The agent for which tools are being executed.
+            agent: The agent (Agent or BidiAgent) for which tools are being executed.
             tool_uses: Metadata and inputs for the tools to be executed.
             tool_results: List of tool results from each tool execution.
             cycle_trace: Trace object for the current event loop cycle.
@@ -78,7 +79,7 @@ class ConcurrentToolExecutor(ToolExecutor):
 
     async def _task(
         self,
-        agent: "Agent",
+        agent: Union["Agent", "BidiAgent"],
         tool_use: ToolUse,
         tool_results: list[ToolResult],
         cycle_trace: Trace,
@@ -93,7 +94,7 @@ class ConcurrentToolExecutor(ToolExecutor):
         """Execute a single tool and put results in the task queue.
 
         Args:
-            agent: The agent executing the tool.
+            agent: The agent (Agent or BidiAgent) executing the tool.
             tool_use: Tool use metadata and inputs.
             tool_results: List of tool results from each tool execution.
             cycle_trace: Trace object for the current event loop cycle.
