+++ src/strands/tools/executors/sequential.py
@@ -1,6 +1,6 @@
 """Sequential tool executor implementation."""
 
-from typing import TYPE_CHECKING, Any, AsyncGenerator
+from typing import TYPE_CHECKING, Any, AsyncGenerator, Union
 
 from typing_extensions import override
 
@@ -11,6 +11,7 @@ from ._executor import ToolExecutor
 
 if TYPE_CHECKING:  # pragma: no cover
     from ...agent import Agent
+    from ...experimental.bidi.agent.agent import BidiAgent
     from ..structured_output._structured_output_context import StructuredOutputContext
 
 
@@ -20,7 +21,7 @@ class SequentialToolExecutor(ToolExecutor):
     @override
     async def _execute(
         self,
-        agent: "Agent",
+        agent: Union["Agent", "BidiAgent"],
         tool_uses: list[ToolUse],
         tool_results: list[ToolResult],
         cycle_trace: Trace,
@@ -33,7 +34,7 @@ class SequentialToolExecutor(ToolExecutor):
         Breaks early if an interrupt is raised by the user.
 
         Args:
-            agent: The agent for which tools are being executed.
+            agent: The agent (Agent or BidiAgent) for which tools are being executed.
             tool_uses: Metadata and inputs for the tools to be executed.
             tool_results: List of tool results from each tool execution.
             cycle_trace: Trace object for the current event loop cycle.
