--- tests/strands/experimental/bidi/hooks/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for BidiAgent hooks."""
